Use TypeNameResolver in ILCompiler.Trimming.Tests
