+++ src/coreclr/tools/aot/ILCompiler.Compiler/Compiler/Logging/DocumentationSignatureGenerator.PartVisitor.cs
@@ -8,8 +8,7 @@ using Internal.TypeSystem;
 
 namespace ILCompiler.Logging
 {
-
-    public sealed partial class DocumentationSignatureGenerator
+    internal sealed partial class DocumentationSignatureGenerator
     {
         /// <summary>
         ///  A visitor that generates the part of the documentation comment after the initial type
